rework owning constructors to be more straightforward
Owning constructors can take advantage of the ownership to fix some
problematic invariants, ie the indices can be sorted by the
constructors. Given this fact, it's better to have a constructor that
panics on bad input (out of bounds indices, etc) since the remaining
structure checks are there only to prevent programming errors.