Experiment with slight acceleration of smmp using a bool vector
Using this bool vector enables to directly store the column indices in
the result indices vector, and we should get less cache misses thanks to
the usage of a boolean vector to check for duplicate column indices.

I also introduce an API to dynamically choose the strategy, to be able
to benchmark them. It's possible I keep this API in the future, if it
turns out there are some good strategies for specialized workloads.

Future work: try a bitvector instead of a bool vector. Try and sort the
indices on insertion.