Fix permutation bug in sprs-ldl
As noted in the documentation of transform_mat_papt, the same process
works whether the input matrix is CSC or CSR.

This is a breaking change, but this API is kept unstable as it is very
prone and its use case is quite specific to `sprs-ldl` for now. We now
enforce this more by hiding this function from the docs.