Accelerate ldl by removing bound checks in hot loops
One bound check could be removed by leveraging iterators, but the other
required the use of unsafe code. It is however quite safe as the indices
are always written in the function in question, and a single assert at
the start of the function can check if the indices will be in bound.