Add ergonomic middle_slice functions for `CsMatBase` variants
To be able to extend the lifetime of the `middle_slice` method targeting
`CsMatViewI` without having to use an other name, I'm experimenting a
new approach, where I re-implement the method for each main variant. Of
course I'd still like to be able to write a single implementation, but I
guess only a future version of rust, with specialization, or maybe
associated type constructors, will be able to help.