Split the error type between structural and linalg errors
Recent work has made the structural constraints of sparse matrices and
vectors clearer, making it clear that there can be a single enum
discriminating all possible structural errors. Having linalg errors in
the same error type was clumsy: structural errors happen when creating a
matrix, while linalg errors happen much later, when solving systems or
performing other linear algebra operations. Therefore, by splitting the
types, it should be easier to react to both kinds of errors.

Since the linalg part of this crate is still small, it's expected that
I'm missing possible errors, which is why the error type is a non
exhaustive enum.

Additionaly, I took this refactoring as an opportunity to fix #255,
using the same pattern used by the `CsMatBase` and `CsVecBase`
constructors to return the storage in case of an error.

As [suggested by
@mulimoen](https://github.com/vbarrielle/sprs/pull/268#issuecomment-755531368),
the `SprsError` type is now an enum around all possible errors raised in
the crate, which can prove useful for some cases.

Co-authored-by: Magnus Ulimoen <magnus@ulimoen.dev>