Support generic dense vectors in more APIs
Lots of APIs were using slices as their input, particularly when using
output buffers. However, this was not a good fit for linear algebra,
since consumers of the APIs are more likely to be using eg an array from
ndarray.

Here we extend the `DenseVector` trait with a mutable version to be able
to use it on output parameters. Since these are sealed traits we should
be able to add unsafe indexing if necessary without a breaking change.
It should also be possible to support eg nalgebra in the future without
too much trouble.

This should improve the situation discussed in #93, though it's
probably not done yet.

As suggested by @mulimoen, the `index` and `index_mut` implementations
hint as `#[inline(always)]` as we want them to be zero-cost
abstractions, and it can be critical to have them inlined to allow the
compiler to remove bounds checks when possible.